[ovsp4rt] Move session files to their own folder
- Moved ovsp4rt_credentials and ovsp4rt_session to a 'sessions'
  subdirectory.

- Made a couple of unrelated maintenance edits.
