+++ ovs-p4rt/sidecar/ovsp4rt.cc
@@ -10,9 +10,9 @@
 #include "logging/ovsp4rt_logging.h"
 #include "logging/ovsp4rt_logutils.h"
 #include "ovsp4rt/ovs-p4rt.h"
-#include "ovsp4rt_credentials.h"
 #include "ovsp4rt_private.h"
-#include "ovsp4rt_session.h"
+#include "session/ovsp4rt_credentials.h"
+#include "session/ovsp4rt_session.h"
 
 #if defined(DPDK_TARGET)
 #include "dpdk/p4_name_mapping.h"
